Get 3 ips regardless of the number of AZ
If there are 3 AZs (and therefore 3 subnets) then no problem
If, however, there are 2 AZs then using an index greater than the length
of the list it will wrap from the beginning[1]

The aim for this is to keep the code simple and not having to use
cumbersome hacks based on `count`.

TBD if it works.

1: https://www.terraform.io/docs/configuration/interpolation.html#element-list-index-