Interpolate HAProxy task def vars & add linux caps
We have to run HAProxy on port 443 and port 80 in order to get static
IPs on our ALB:

static EIP NLB -> HAProxy -> ALB

NLBs are not smart enough to do proper port routing so the listener and
HAProxy have to listen on the same port.

In commit 588e6dff1e3945cdd00118c555661e590c1557cd we added env vars to
configure HAProxy:

- BACKEND (where the traffic goes)
- PORT (that HAProxy listens on)
- RESOLVER (where HAProxy goes for DNS)

- BACKEND will be set in the task definition template_file data resource
- PORT will be set in the task definitionm template_file data resource
- RESOLVER is set to the magic AWS IP for DNS (google it)

This commit also adds the extremely esoteric options that we need to
listen on 0.0.0.0 (inside a container) and on privileged ports.

The magic capabilities that I found (through `man capabilities` and by
`docker run`) were:

- Add CAP_NET_ADMIN capability (interfaces)
- Add CAP_NET_BIND_SERVICE capability (privileged ports)

Additionally I needed the `net.ipv4.ip_unprivileged_port_start` sysctl
because otherwise the capabilities were not enough. Perhaps this option
is enough and the capabilities are superfluous but that seemed not to
work.