Move the ALB to the internal subnets
VPC NETWORKING INCOMING

Our network topology for static IPs for metadata is as follows:

static EIP NLB -> HAProxy -> ALB (TLS termination)

Our NLB has to be external in order to receive traffic from the
internet.

The NLB will forward traffic to the HAProxy containers running on EC2
instances. If these instances are going to be able to connect to the ALB
then there are two options: (in the case where the ALB is public)

- The boxes have public ip addresses (they would have public IP
addresses but the connection to the ALB (external) would be done inside
the VPC due to some absolutely confusing AWS magic) - this in my opinion
is far too confusing

- The boxes could talk to the ALB by following the routing table all the
way from their subnet, through the NAT gateway, over the public
internet, into the public ALB.

I thought both of these were suboptimal, so moved the ALB to be
internal. This means that HAProxy can connect directly to the ALB inside
the VPC, and the network topology is a lot simpler.