Split static_ingress listener into two (http/https)
ECS only supports (currently) a single target group and load balancer
per ECS service. This means we have to run two services if we are going
to run HAProxy on port 80 and port 443.

We run on both port 80 and 443 in order to get the ALB to redirect
traffic to get the 301 on HTTP to HTTPS.

This commit splits the target group and listeners into an `_http` and
`_https` which listen/forward on/via port 80 and port 443 respectively.