REVERT ME LATER: reduce number of targets for NLB
We have an NLB for static ingress

NLBs have subnet_mapping configurations defined on the resource, which
makes it impossible to make them aware of how many AZs we have. So we
hardcode three targets and IPs, however in environments where we have
only 2 AZs there are three IPs, one AZ has two IPs.

However this commit hardcodes it to two EIPs and two subnet mappings as
a result. This is because in Joint, which is our only presently built
environment, AWS has limited us to 12 EIPs which we have reached.

I have raised a support ticket but in order to get terraform which
applies we have reduced the amount of EIPs for this deployment.

We should revert this commit later