Resolve metadata internally
As of commit b5f0dc9a2752cf3949bbce52db92ebd211675bae the ingress ALB
which serves www, metadata, and analytics, is internal, which means we
don't talk out to it anymore.

We need to let things that can connect to metadata connect to ingress in
order to for the hub to work.

This allows services to connect to it with a security group rule and
updates nonProxyHosts