Increase number of hub apps to len(AZs)
These were 1 before whilst we were testing and didn't have kibana.
Also we were limited to only running one policy due to sessions in
infinispan (boo).

Now we are using redis for sessions in policy this is not the case.
Also logs go to logit now so we aren't on the box to check so this is no
longer a usability problem.

This is good because Prometheus will no longer permanently have noisy
DOWN targets, and is an opportunity to test our load balancing.