Change ECS rolling update params to (50/200)
This adds optional vars for our ecs_app module which control ECS rolling
update.

ECS rolling update has 3 parameters it looks at:

- number of targets
- maximum %
- minimum healthy percentage

We need minimum healthy percentage (for hub apps) to be 50% so it will
remove one app when doing new deploys if everything is healthy,
otherwise we will not have an instance to schedule the new app on.

200% is the default for maximum healthy percentage