Add ECS vpc endpoints
This will allow us to communicate with the ECS APIs from within our VPC,
not over the private internet. This creates the VPC endpoint but doesn't
get the boxes to use it.

We will need to enable connectivity to the security group.

I'm reusing and renaming the ecr security group because we won't have
one without the other.