Deploy prometheus using ECS
We would like to have a more recent version of Prometheus than is
available in the ubuntu packages, but we don't want to set up a debian
packages pipeline.

We can deploy prometheus using ECS as an ECS daemon; ECS will put a
prometheus container on each host instead of us using systemd to keep it
running.

The task has to run with network_mode = "host" so that it can scrape
other containers using localhost instead of docker networking

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>