add beat-exporter daemon service to each cluster
This adds beat-exporter to each cluster, so that we can export
journalbeat metrics in prometheus format.

We have lots of different ECS clusters and there isn't a nice way of
running the same service in every cluster, so I've had to hardcode a
list of clusters and manually construct the ARNs for those clusters.