Journalbeat needs to be restarted frequently.
Journalbeat appears to have a fault that causes it to stop delivering
logs. This is hard to detect since it does not error and also remains
running. The simple, quick and dirty fix is to restart the service on a
regular basis.

Logs back up and are not lost but are resent after the service has been
restarted.

Single: matthewcullum-gds