set journalbeat alert threshold to 0.01
The `rate()` function returns a rate-per-second.  We should tolerate
quite a low rate.  I think we can expect a minimum of:

 - 3 log messages per cron job
 - 1 cron job every 5 minutes

for a total of 9 messages over a 15 minute interval.  This corresponds
with a nice round 0.01 messages per second.  So this commit says alert
if we're sending less than this number.