implement alerts in prometheus
This creates an alerts.yml file which defines alerts for prometheus to
fire.  It sets up just 2 alerts for the moment: Watchdog, which will
go to cronitor, and HubSamlProxyErrorsReceivingRequest, which is one
of our page-type alerts.

Currently, they don't go anywhere, but future work will connect
prometheus to an alertmanager.

They use the same base64-encoded environment variable trick as the
main prometheus.yml file.