Add haproxy-static-ingress-tls image
Prior to this commit our assumptions about the hub's infrastructure
were:

- static ingress
- static egress
- tls everywhere

Recent events have made us realise that having only these assumptions
would cause a regression. The new assumptions are as follows:

- static ingress
- static egress
- tls everywhere
- source ip of the user is present in tls header

The last assumption is key to how the billing and audit subsystem works,
as well as how our analytics subsystem works.

In order to do this we have to change the architecture of the hub
somewhat. We must now terminate TLS at the NLB. This is fine because we
already have an appropriate ACM cert for doing this.

The old flow was

NLB (TLS) ->
  HAProxy (TLS) ->
    ALB (TLS re-encryption) ->
      app (nginx/ms)

The new flow is

NLB (ACM TLS re-encryption) ->
  HAProxy (TLS re-encryption) ->
    ALB (ACM TLS re-encryption) ->
      app (nginx/ms)

This involves an extra TLS re-encryption step but this is "fine" because
verify values security over latency for the user.

This commit is the first step in a few to make this happen and
introduces a container image for HAProxy for HTTPS -> HTTPS.

The container starts up and generates a new key & self-signed cert
This is concatenated into a format usable by HAProxy

The log level is set to info instead of warning because HTTP log is
valuable if slightly verbose

The Alpine ca-certificates package is installed so we can do TLS
validation against the ACM cert on the ALB.

The reasons why I have chosen haproxy:

- Nginx has issues with DNS resolution, and you must do the `set $` hack
- Haproxy has more explicit and better DNS resolution
- I do not trust nginx to not do something stupid with the http request
- We were using haproxy before (just at a different layer)

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>