BAU: Update nginx to log message in JSON format by default
This will enable journalbeat to be able to decode nginx JSON log message
to make it easier to see what is going on in Kibana. The nginx json log
format is taken from nginx config file in old infrastructure. This will
keep the nginx log format same between old and new infrastructure.

Author: @adityapahuja