Ensure system logs are not duplicated
We rely on journalbeat to ship the systemd journal to logit.

The default behaviour of systemd-journald is to also write logs to
syslog, this results in duplicated logs that fill up the disk.

We noticed this behaviour post-switchover because of real traffic
volumes.

This commit turns off the default behaviour of writing the journal to
Wall and to Sylog. It also makes it explicit that we should use a
maximum of 2G of disk.

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>

Co-authored-by: Dean Wilson <dean.wilson@digital.cabinet-office.gov.uk>
Co-authored-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>