Add explicit prevent_destroy to aws_eip resources
We have Deny on destroying EIPs in our IAM policies, however Concourse
does not.

If someone did a bad then Concourse might be able to destroy the EIPs,
our tests should catch this in joint but then cleanup would be a pain.

This commit adds the prevent_destroy annotation which will make
terraform reject any config which accidentally destroys this

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>