Add new aggregate Saml Proxy alert
This replaces the 4 existing alerts with 1 which has the same logic
but the 95% success threshold is checked against the total number of
requests.

The previous solution was very trigger-happy in situations where the
traffic was low (nights, weekends, or both) and even a single 500
error would make the success rate drop below 95% and result in a
pagerduty alert.