split infra and service ticket dashboards
This splits the ticket-type alerts dashboard into service-level alerts
and infrastructure-level alerts.  The service-level alerts are:
certificate expiry, metadata expiry, OCSP alerts.

The infra-level alerts are things like NTP, disk space, prometheus
scrape failures, journalbeat.

The service-level dashboard has the "Verify" tag, and the infra one
has the "RE" tag.  This allows us to show only service-level tickets
on one dashboard and only infra-level tickets on another.

While I did this, I also changed the "on error" state - for all alerts
this was previously set to "Alerting".  This meant that we had a lot
of noise for when grafana had temporary network blips talking to
prometheus.

I have changed this so that only the Prometheus targets alert will
alert if it can't hit prometheus.  This is a compromise so that we
still have visibility if grafana can't access prometheus, but we don't
fire every alert for every random network blip.