Allow prometheus to scrape frontend
This commit allows prometheus to scrape frontend metrics (which were
added in alphagov/verify-frontend#674).

This is a bit different from how we scrape the other apps.  Other apps
expose their ports as ports on the host, but frontend uses `awsvpc`
networking, which means that the frontend task gets its own Elastic
Network Interface (ENI), and so is accessed on a separate IP address
from the host itself.

This means that we can't use prometheus's ec2_sd_configs to find
frontend apps like we do for the dropwizard apps.  Instead, we use AWS
Service Discovery.  Service Discovery is a powerful set of tools but
the bits we care about are:

 - we create a private DNS zone for discovering hub apps. This is the
   aws_service_discovery_private_dns_namespace.
 - we create a "service" to represent the frontend instances.  This is
   the aws_service_discovery_service.  You can register instances into
   this service manually if you like, and they will appear in DNS. Howver:
 - we configure the frontend ECS service to automatically register
   tasks as instances of the service discovery service.

We use SRV records for this, which prometheus supports by default in
its `dns_sd_configs` configuration block.  This means that prometheus
can look at the SRV record at `frontend.hub.local` and get a list of
service instance records, which give a hostname and port in the form:
210a00f0-5c70-4a6e-87e9-db440c7dab51.frontend.hub.local:8443

Finally, we need to open the security groups so that prometheus has
permission to access frontend.

I've tested most of this by manually applying to joint, and verified
that the ECS service registers itself and prometheus can discover the
instances.  The one bit I haven't tested is the security group rule -
but I think this PR is good to merge and see if it works rather than
continuing to mess around outside the pipeline.