Use the JournalBeat OSS package
- In 67aeed1113625628c912ccd5da623184f72aaa71, we upgraded to
JournalBeat 6.7.0. When the Prometheis respun, we got these errors in
the JournalBeat logs:

```
Connection marked as failed because the onConnect callback failed: This Beat requires the default distribution of Elasticsearch. Please upgrade to the default distribution of Elasticsearch from elastic.co, or downgrade to the oss-only distribution of beats"}
```

- Given we have no control over the ElasticSearch distribution that
LogIt uses, let's use the OSS package. It fixes the error (we tested on
one instance of Prometheus in Joint).