Port grafana ticket alerts to prometheus
These alerts currently exist on the grafana infra ticket dashboard but
they all go into verify 2ndline slack, causing noise.  Also the
TargetDown alert doesn't work well in grafana, because currently it
alerts if any instances have been down for a period of 15 minutes,
even if no single instance was down for that long.

It's worth re-documenting some of the alerts here as well:

NtpOffsetTooGreat was ported over from the old infrastructure.  We
might be able to do a better alert based on the node_ntp_sanity
metric.

DiskPredictedToFill was noisy when we first created it.  The
node_creation_time guard means we only alert for instances that have
existed for a day or more, because for instances younger than that it
was very trigger-happy.

DO NOT MERGE BEFORE
https://github.com/alphagov/prometheus-aws-configuration-beta/pull/292
or you will ring the observe pagerduty a lot