Make deployment maximum percent = 100 (everywhere)
`deployment_maximum_percent` determines what percentage of
`desired_count` ECS will attempt to schedule. Because we pin one app to
one EC2 instance we do not really have capacity for > 1 app per EC2
instance.

Explicitly setting 100% as the max will ensure ECS gets rid of a single
old app and schedule a new one. This was done for static-ingress tasks
in acacb08 and has reduced the amount of time taken to deploy changes to
static-ingress

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>