reduce the number of CloudWatch metrics we collect
We're spending quite a bit on GetMetricsStatistics API calls.  We can
reduce that by requesting fewer metrics.

For this commit, I focussed on retaining only those metrics required
to support existing alerts.

We don't use the RequestCount or RequestCountPerTarget or
TargetResponseTime metrics in any alerts, so they can go.

We use the Target_[2345]xx_Count metrics, but our alerting rules
aggregate away the AvailabilityZone label.  So we can just not request
it in the first place and reduce our bill on these metrics by a factor
of 3.