Reboot at instance creation time in order to pick up new kernels
Canonical do not bake new AMIs as frequently as they release kernel
patches.  Therefore, when we launch an instance, it may not have the
latest kernel.  This causes alert noise because we see lots of
RebootRequired alerts.

We need to ensure that everything we want to run will survive the
reboot.  We run most things as systemd services, so that will be
fine.  The ECS agent is run via docker.  If we set `--restart=always`,
then the docker daemon will restart the ECS agent on daemon start --
even after a reboot.