Really really upgrade kernels when we provision instances
`apt-get upgrade` won't install new kernels because it involves
installing new packages.

`apt-get dist-upgrade` will.

I considered that `dist-upgrade` is a bit of a sledgehammer and it
will also *remove* packages, which we don't want, but the risk of this
actually a) happening, and b) breaking something, feels relatively
small on a freshly-minted machine from an official AMI.