start journalbeat after docker
We observed that journalbeat wouldn't come up after a reboot.  After
much digging, we suspect that it gets stuck because docker isn't
available.  We manually bodged the systemd unit for journalbeat to
require docker.service and rebooted the instance, and journalbeat came
up normally.  This seems to support our theory.

Sorry-not-sorry for the gratuitous `sed`.

Also, replace separate `systemctl enable` and `systemctl start` with
`systemctl enable --now` :golf: (thanks, @tlwr).

Co-authored-by: Alex Monk <alex.monk@digital.cabinet-office.gov.uk>