Add Self Service app secret key base to Parameter Store
This will now generate a random string for rails secret key and save it
to the Parameter Store.

Also update the self service app task definition to use that generated
secret key.

This saves us from manually generating a key and adding it to the
Parameter Store.

There is currently a rails secret key in the staging account that was
manually created. This will be removed a little later.