Add Self Service Cognito User Pool
The Self Service app will be managing its' users via AWS Cognito. This
adds the config for it.

MFA config is off by default. We will be controlling this flow through
the API via the app so this is fine.

Also add the iam policy for the self service fargate task to talk to
cognito.