Add Integration config metadata bucket arn to Self Service policy
When the Self Service app is in production it will also be reading and
writing to a bucket in Integration.

Once we deploy out to production access will be limited only for the app
running in that environment. For testing right now we're giving to the
app running in staging.