Use temporary validity password attribute in Self Service app
AWS changed their cognito API and the old field `UnusedAccountValidityDays`
is now TemporaryPasswordValidityDays. These translate to snake case
alternatives in Terraform. The attribute now also exists in the password policy block instead of the admin user block.

Unfortunately the official provider may not be updated for a while.

In the mean time we have forked and updated our own AWS provider.
This means we have created our own binary version and will be using
that with the Self Service app until the official update happens.

We will be adding the use of the custom provider to the Self Service
deployment pipeline.

[This corresponding infrastructure config PR](https://github.com/alphagov/verify-infrastructure-config/pull/202)
will enforce the use of the custom provider.

Co-authored-by: Jakub Miarka jakub.miarka@digital.cabinet-office.gov.uk