zfVNkzhm: Allow suppling the db command
The way terraform is currently set up means that the DB migration task shares the command across all the environments. Meaning that if we wanted to for example drop a database in production, it will drop it in all the environments as it goes through the pipeline.

We should be able to specify the command per-environment basis. It's essential especially when spinning up a new environment where we want to run the `db:setup` command.

This adds a new variable with a default of `migrate`. It will now allow us to
supply a different command per deployment.