Bump the RDS Postgres engine version and RDS root CA
It doesn't appear that the `self-service` Rails app is verifying the
certificate RDS presents so we should be good to just start presenting
the new one.

We're also bumping the version of Postgres to match what is currently in
production. For some reason Postgres in production is version 11.5 where
as staging is 10.6.