AUT-606: Change ECS scheduler's placement strategy for ingress cluster
At this moment, ECS scheduler is using random for ECS task placement strategy. This issue with the strategy is that it places all same containers (e.g. metadata) on one EC2 instance. This is not a good strategy and it increases the risk of losing a service if all same containers are on one EC2 instance. This change updates ECS scheduler to use spread for ECS task placement strategy. This strategy will ensure that tasks are distributed evenly across all instances. This change applies to metadata, analytics and frontend containers. beat-exporter's service type is daemon and it can be safely ignored.

Source: https://docs.aws.amazon.com/AmazonECS/latest/developerguide/task-placement-strategies.html

Author: @adityapahuja