Detect device identifier of prometheus EBS volume for mounting.
EBS volumes are mapped to device identifiers in the kernel in a way that
is unpredictable (it gives numbers and letters based on the order in which
it becomes aware of them and so it cannot be relied upon). We are seeing
an issue whereby the ordering of the devices sometimes reverses and so an
attempt is made to mount the wrong device.

Also, the check for whether the disk was formatted was failing sometimes.
The `file -s` command wasn't returning the expected string and so a disk
that needed formatting wasn't being formatted. The mount command that
follows then fails because there's no filesystem to mount.

This change attempts to discover the identifier by looking it up by disk
size. It's a hack, but, for the moment, prometheus is only using a single
disk of that size. The failing `file -s` check has been replaced by a
check interrogating `blkid` which seems to be more reliable in testing.