Ensure the right process gets the signals.
A shell won't necessarily propagate an incoming signal to its child
processes. In the cases of these Verify components there's no need for the
shell itself to continue to exist so the relevant application can become
"pid 1" and take over. This brings these containers inline with current
best practices in docker deployments.