Try running apt forever in ECS cloud-init
We keep seeing issues during EC2 instance cloud-init where APT fails to
run successfully because it can't access a lock it needs. Instead of
checking for the locks ourselves and still hitting a potential race
condition, swtich the logic so we just keep retrying on failure. The end
result may be the same (an EC2 instance that fails to join the ECS
cluster) but this way there's more chance of it eventually succeeding.