configurable min/max asg sizes
allow configuring min_size and max_size of ASGs setup via the asg
module.

this is to enable work to move to capacity provider congiured asg sizes.

the desired_capacity config is dropped as terraform defaults this to the
min_size if not present (which is same as before) and keeping this value
set may conflict with capacity provider