Improve stability of docker state during ECS EC2 initialisation.
We occasionally see issues where tasks can't start on an instance. The port
mapping required by the task can't be satisfied as the port is already
bound on the host. The working hypothesis is that this is due to docker
getting into a bad state as a result of the reboot at the end of the
cloud-init script leaving port mappings behind. These port mappings get
restored on reboot by docker before any tasks are (re-)registered by ECS.

By moving control of the ECS daemon to systemd, and not running ecs during
cloud-init we should see a reduction in the recurrence of the docker
state problem.