let ECS manage the ingress ASG size
the ASG capacity provider lets ECS decide the required size of the ASG
based on the tasks running on the cluster.

this configures a capacity provider for the ingress cluster so that we
can decouple the size of the cluster from the desired number of
frontend, analytics and metadata tasks we want to run on it.