"spread" placement for ingress services
we would like for instances/tasks of the same service to execute
seperate EC2 instances to maintain high availability at the cost of less
effcient binpacking.

this is more important now that capacity providers manage the size of
the ECS cluster as we want it to prefur scaling up to more instances
than packing more tasks onto an already busy node.