Add saml-engine fargate app
This (hopefully) adds a saml-engine app in fargate.  Nothing is hooked
up to it yet.

I've done a test terraform plan, it adds 26 resources, changes 0 and
destroys 0.

I tried to avoid duplicating the task definition file so we don't have
two parallel task definitions.  To do this I added an `optional_links`
parameter which can be removed when we're not running saml-engine in
EC2 any more.

For the new app I used the new terraform `tempatefile()` function in
favour of the now-deprecated `data.template_file` data source.