Add generic hub_fargate_microservice security group
It feels like a lot of duplication that prometheus has to be allowed
to specifically talk to each hub app individually.  This creates a new
security group, hub_fargate_microservice, which is intended to be
added as an additional_task_security_group_ids to each fargate app.
It allows prometheus to talk to 8443 on any app that has this security
group.  Longer term, we will only need 1 SG rule where previously we
needed 5 or 6.