analytics: add analytics task to fargate cluster
deploy second set of analytics tasks into the shared fargate ecs
cluster.

the tasks are connected to the existing target group so will serve
traffic in parallel with the existing EC2 tasks.

Due to fargate cpu/memory constraints the task definition cpu requests
are lowered a little to fit.

The memory requests are left as is for now as they are tuned to fit a
certain number of tasks onto the EC2 instances. A follow up PR will
alter these values to use the maximum memory for the task once
EC2-service is removed.

Service discovery doesn't do much for these analytics instances, it is
mainly here for:

1. consistancy with the other tasks
1. in case we need it in the future, adding it now avoids the need to
   recreate the ecs_service later.